Safari-611.1.10/LayoutTests/crypto/subtle/rsa-import-jwk-key-export-jwk-key-public-expected.txt

Test importing a JWK RSA-OAEP public key with SHA-1 and then export it in JWK format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
Exporting a key...
PASS exportedJwkKey.kty is jwkKey.kty
PASS exportedJwkKey.alg is jwkKey.alg
PASS exportedJwkKey.key_ops is jwkKey.key_ops
PASS exportedJwkKey.ext is jwkKey.ext
PASS exportedJwkKey.n is jwkKey.n
PASS exportedJwkKey.e is jwkKey.e
PASS successfullyParsed is true

TEST COMPLETE
